update base class and extend it for registration flow
